+++ src/Fable.Cli/CHANGELOG.md
@@ -9,6 +9,7 @@ and this project adheres to [Semantic Versioning](https://semver.org/spec/v2.0.0
 
 ### Added
 
+* [Python] Support for Nullable Reference Types (by @dbrattli)
 * [JS/TS] Fix #3533: Add directives prologues supports (by @MangelMaxime)
 * [JS/TS] Support for Nullable Reference Types (by @ncave and @MangelMaxime)